RuleDefinitions tests, inclusion of jest-yaml.
