+++ client/package.json
@@ -150,6 +150,7 @@
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^4.2.2",
+    "yaml-jest": "^1.0.5",
     "yaml-loader": "^0.6.0"
   }
 }
